# feat: improve globbing performance by using parallelized Swift glob implementation