--- a/Package.resolved
+++ b/Package.resolved
@@ -32,8 +32,8 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/tuist/swift-glob",
       "state" : {
-        "revision" : "c2bf1c8cce6c2ee91bd8ed57080ff00acfb6f3fb",
-        "version" : "0.3.3"
+        "revision" : "a32f83e851593b4bec436ebac910d70af6d77f6a",
+        "version" : "0.3.4"
       }
     },
     {