… Sources/FileSystem/AbsolutePath+Extras.swift
… Sources/FileSystem/FileManager+Extras.swift
… Sources/FileSystem/Glob.swift