We have recently [moved away](https://github.com/tuist/FileSystem/pull/66) from the `swift-glob` library due to a couple of issues with the library's implementation.

I fixed all of those issues [in a fork](https://github.com/tuist/swift-glob/pull/1) with the hope that I'll find the time to push those changes upstream.

Moving to `swift-glob` improves our generation times by up to 20 % from the current state of `main` in `tuist/tuist`:
```
Benchmark 1: "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open
  Time (mean ± σ):      4.503 s ±  0.132 s    [User: 9.485 s, System: 3.667 s]
  Range (min … max):    4.388 s …  4.827 s    10 runs
```

Compared to numbers from the latest [perf improvement](https://github.com/tuist/tuist/pull/6915):
```
# Warm generation runs for a real-world project
Benchmark 1:  "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --path /Users/marekfort/Developer/nextdoor/iOS --no-open
  Time (mean ± σ):      5.419 s ±  0.142 s    [User: 9.960 s, System: 3.771 s]
  Range (min … max):    5.304 s …  5.803 s    10 runs
```

And compared to the initial baseline, we have already made ~ 30 % performance improvements:
```
# Warm runs
Benchmark 1: tuist generate --no-open
  Time (mean ± σ):      6.354 s ±  0.244 s    [User: 10.129 s, System: 2.629 s]
  Range (min … max):    6.096 s …  6.816 s    10 runs
```